Create new classes Ui, Parser
Rename FileHandling class to Storage
