+++ src/main/java/melchizedek/Parser.java
@@ -1,0 +1,4 @@
+package PACKAGE_NAME.melchizedek;
+
+public class Parser {
+}
